requirements.txt
streamlit
openai-whisper
torch
groq
keybert
transformers
nltk
matplotlib